Fix wrong identify index of array loop
Currently riotjs uses an array of difference between the mounted data and recent updated data and sets the index from this difference. 
When you define a loop, waits that the index is the current matrix and not the your difference.

@example
<todo each={ obj, index in items }>...

Currently index is a index of a difference. This patch use a pos value returned by indexOf.