Add a new simple tutorial, how to use babel and riot to build app.
This example is slighly different witht the existing one with webpack.
This one uses browserify and it uses Riot's tag files.
This one uses cli to build the app, not gulp/grunt.
This one also covers how to use co and generators in the app.

Also, remove some dup entries in the READMD.md