tmpl.js complete rewriting
@tipiirai @GianlucaGuarini @cognitom @rsbondi , hi

Last days I spent some time cleaning and testing tmpl.js, and reviewing its logic.
This is the result.

I wanted above all: backward compatibility, correct details that have generated bugs & issues,
provide flexibility in the implementation, and facilities to understand the code through multiple comments; all without the code grew ...too
(the minified size of this version is XXX bytes vs XXX bytes of the one in the master branch).

But I have not had time to study other parts of riot, and how they interact with tmpl,
so perhaps tmpl caller filters out characteristics that this code implements, and
therefore has grown unnecessarily.
EDIT: In testing with some issues, I think there are details in the compiler that cause these issues. The string received from tmpl is filtered, new characteristics are not so usefull, and in other cases, the output of tmpl is ignored.

This file passed all riot tests and, in my opinion, is in beta phase and ready for testing on
performance and in real environments, so please, break the code now.
I'm test it in latest IE, FireFox, and Chrome. I will do some mobile tests later.

Code is heavily commented 'cause today, we know what we doing, tomorrow... God knows?
(Corrections are welcomed, the English is not my friend)

#784 is fixed
#744 is related to the compiler

Thanks.