Prevent loop state from leaking out and getting modified.
`each` has it's own array that tracks the tags for every item. This array was being assigned to the parent's `tags` collection, thus leaking outside of `each`.

If another instance of that tag is created outside the loop, it gets added to that array (by addChildTag). Now, when `each` wants to remove tags (because items have changed) it will inadvertently remove the instance that was created outside the loop.

Instead of just setting parent.tags, use the addChildTag and removeChildTag to modify the parent's collection. This has the downside that parent.tags order won't match DOM order, but it's much less code.