Don't update expressions or tags within a false if-attribute.
Before, expressions and child tags were run even if they never made it into the DOM. This was unnecessary work, and sometimes caused spurious errors.

The change combines parsing into a single function that finds all expressions and child tags. The parser returns a tree, so we know what expressions are under what if-attribute.

Instead of using observable, updates propagate through this tree, which lets us stop at a false if-attribute. Child tags are only created and mounted if they're actually going to be inserted into the DOM.

The most notable side-effect of this change is that child tags are available later than they used to be. Before, it was one of the first things that happened, so you could reference them in the init function, mount, or update callbacks. Now, they aren't available until after the first update has finished.