Refactor how tags get opts from their parents.
Before this change, attribute expressions on a tag were evaluated twice: once by parent as part of the standard update, and then again by the tag itself inside of updateOpts. The first one actually modified the DOM, and the second one only affected the opts.

This was wasteful, and we should really only evaluate each expression once per update cycle. Here's how it works:

When parsing, we keep track of every attribute, whether it has an expression or not. These are stored as ownAttrs on the tag. The attributes with expressions are also stored on the parent's expression tree. When an update happens, the parent evaluates all its expressions (include the attributes on the child tag), then calls update on the child. Our updateOpts method now just copies the values of those already-eval'd expressions onto opts.

This also lets me revert the one dubious change I made to the tests. In the last version attributes got eval'd by the child tag first, and in the case of callbacks, that actually modifies the DOM and prevents the parent for eval'ing onmouseenter.