Unmount things when an if-attribute becomes false.
Before this change, we'd keep the old DOM, and simply avoid updating if when the if-attribute was false. It's more predictable to say that when DOM goes away, it's like it was never there to begin with. When the if-attribute becomes truthy again later, we should insert and mount new, clean DOM. This goes for custom tags as well as built-in stateful tags (like inputs).

To do it, I made if-attributes work similar to each. When parsed, it takes that element (plus all descendants), removes it from the DOM, and stores it as the pristine copy.

On update, if the value is true, we clone that pristine DOM, parse, and update it.

This change is breaking.