Add support for debugging tests via chrome.
This lets us use the chrome dev tools to debug failing tests. Karma gets unhappy when files disappear, so I removed `clean` from `make raw`. This way, you can change the source, run `make raw` and reload the browser.

I also disabled the coverage preprocessor in this mode, because it obfuscates the code.

This is the setup I've been using for a while. It isn't perfect, and sometimes karma stops picking up changes to the source, requiring a restart.