Prevent each's internal state from leaking out and being modified.
`each` has an internal array that stores the tag instance of every item being rendered. To make these tags accessible from the parent (eg `parent.tags['child-tagName']`) someone assigned that internal array directly onto the parent.

Problem is, if that parent has other instances of that child tag, they'll start to conflict. If you have two loops over the same child element, they'll end up remove each others items.

This change uses the add/remove methods to modify tags instead. This was easy, but it means that parent.tags['child-tagName'] is less likely to reflect the actual order of elements in the DOM, and I  changed some tests to reflect this.

This is a breaking change (sorta).