Fix each to work inside an if statement.
When the `if` is initially false, then each will miss the before-mount event. Instead, we need to remove the template dom immediately while parsing.

There was also a line that ensured the root pointed to the element that will eventually contain each item. I figured it was easier to just set this every time during update.