Always run update when mounting a tag.
It was inconsistent that update would only sometimes be run during a mount. This caused tags to not update when they were inside an if statement, because self.parent existed.

Instead, always run the update during mount. We also should only set DOM._tag and Tag.isMounted during the mount method, because those are used in other places to identify that the tag has mounted.