Create anonymous, inline tags within tests.
Writing and debugging tests is cumbersome because the tag html is usually in another file, requiring you to switch back and forth.

This change lets you define tags directly in the test, which should be a readability win. It's also my hope that this will lead to shorter, more focused tests...because right now we have enormous ones that are hard to debug.

It uses ES6 template strings, and for that I had to use babel to pre-process the test files.