Be more tolerant of spacing to inline test tags
The compiler expects the closing tag to not have any leading whitespace. That's a bit awkward looking with template strings in a test, so clean it up with a regex.

I also discovered that there was an <inner> tag getting shared between multiple tests, which was very un-intuitive, so I inlined them.