Implements #1319
When an event triggers, we want to traverse the DOM upwards until we find a riot tag, then call update on it.

I think that's the intent behind getImmediateCustomParent, but it didn't skip anonymous tags used by `each`. I fixed that behavior, which simplified the event handler code.

This is a breaking change, and there was one test depending on the old behavior that I had to change.