Invert the logic from `hasImpl` to `anonymous`
This is a functional no-op, since !hasImpl == anonymous, but it makes the code clearer, since the only time hasImpl is false is from anonymous tags created by `each`.

It also works around a bug: when tags were mounted through riot.mount, `hasImpl` didn't get set properly, and was false. By inverting the logic, it defaults to being a regular tag.