Fix: double backslashes on riot path when doing server-side rendering
Server-side rendering injects a require to riot path in tags. The use of
path.resolve() to get the path uses single backslashes as path separator
on Windows, which breaks when trying to
require('c:\whatever\the\path\to\riot').

This fix doubles all backslashes in the path returned by path.resolve().