Relative Pathing and checking for require.extensions
In lib/server/index.js, we use path.resolve(__dirname, '..', '..', 'riot') to access the relative path of riot.js . path.join('..', '..', 'riot') seems more appropriate for a relative path inside a module.

Additionally in lib/server/index.js , we attempt to modify require.extensions[.tag], however, that is not ideal both because someone might have already replqced require (eg: SystemJS) and/or because require.extensions is deprecated. For now, putting an if (require.extensions) around the code at least guards us from a runtime exception, although this may not be the ideal solution because we don't have a replacement for the fact that .tag files won't be able to be required like this. But maybe that's okay because if you are in this problem state, you are using something like Webpack or JSPM which has their own loader for .tag files?